fix: moved findChildrenByKey function to utils
