+++ components/lib/utils/Utils.js
@@ -7,5 +7,6 @@ import { mergeProps } from './MergeProps';
 import ObjectUtils from './ObjectUtils';
 import UniqueComponentId from './UniqueComponentId';
 import { ZIndexUtils } from './ZIndexUtils';
+import { findChildrenByKey } from './FindChildrenByKey';
 
-export { DomHandler, EventBus, IconUtils, ObjectUtils, UniqueComponentId, ZIndexUtils, classNames, mask, mergeProps };
+export { DomHandler, EventBus, IconUtils, ObjectUtils, UniqueComponentId, ZIndexUtils, classNames, mask, mergeProps, findChildrenByKey };
